update : 장바구니 api 수정, 장바구니 스키마 수정

+++ routers/goods.js
@@ -2,8 +2,30 @@ const express = require("express");
 
 const Goods = require("../schemas/goods");
 const Cart = require("../schemas/cart");
+const authMiddleware = require("../middlewares/auth-middleware");
 
 const router = express.Router();
+
+// 장바구니 조회
+router.get("/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
+
+  const carts = await Cart.find({ userId });
+
+  const goodsIds = carts.map((cart) => {
+    return cart.goodsId;
+  });
+
+  const goods = await Goods.find({ goodsId: goodsIds });
+
+  const result = carts.map((cart) => {
+    return {
+      quantity: cart.quantity,
+      goods: goods.find((item) => item.goodsId === cart.goodsId), // array.find
+    };
+  });
+  res.json({ carts: result });
+});
 
 // 상품 목록 조회, ?category=전자기기
 router.get("/", async (req, res) => {
@@ -42,11 +64,12 @@ router.get("/:goodsId", async (req, res) => {
 });
 
 // 장바구니 상품 추가
-router.post("/:goodsId/cart", async (req, res) => {
+router.post("/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -54,12 +77,12 @@ router.post("/:goodsId/cart", async (req, res) => {
     });
   }
 
-  await Cart.create({ goodsId, quantity });
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: "success" });
 });
 
-// goods 생성
+// 굿즈 등록
 router.post("/", async (req, res) => {
   try {
     const { goodsId, name, thumbnailUrl, category, price } = req.body;
@@ -84,15 +107,17 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.put("/:goodsId/cart", async (req, res) => {
+// 장바구니 수정
+router.put("/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
 
   if (existsCarts.length) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
     );
     res.status(200).json({ success: true });
@@ -101,12 +126,14 @@ router.put("/:goodsId/cart", async (req, res) => {
   }
 });
 
-router.delete("/:goodsId/cart", async (req, res) => {
+// 장바구니 삭제
+router.delete("/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId });
+    await Cart.deleteOne({ userId, goodsId });
     res.json({ result: "success" });
   } else {
     res.status(400).json({ message: "카트에 해당 상품이 없습니다." });
